pyproject.toml
[project]
name = "fde-fitness-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "google-genai>=1.73.1",
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "requests>=2.33.1",
    "streamlit>=1.56.0",
    "uvicorn>=0.46.0",
]
